Tool/model for laying out a merged front panel (#165)
`combined_front_panel.scad` is a model for a configurable merged front panel. The frame width/height are configurable, as are the number of rows & columns of modules and the spacing between modules (can be defined as a "gap" value beyond the minimum possible module spacing, or a center-to-center distance).

There is a matching python script `generate_combined_front_panel.py` for exporting the design as an SVG, for use with a laser cutter or CAM tool.

Also adds some minimal support for routers/endmills with non-trivial kerf - only for the tabs on the front panel - which adds the option of dog bones for the corners and allows for using a tool that is larger than `thickness` while still keeping the vertical mounting position constrained.

Example model rendering for a large (4'x8') layout with 108 modules:
![combined_front_panel_example](https://user-images.githubusercontent.com/414890/124689025-a03be680-de8c-11eb-8ac4-bd20be76314a.png)
![combined_front_panel_example_closeup](https://user-images.githubusercontent.com/414890/124689028-a205aa00-de8c-11eb-9edc-4a0fd12f9914.png)